+++ lib/sitehub/http_headers.rb
@@ -1,4 +1,5 @@
 class SiteHub
+  # TODO: - write spec directly against this class
   class HttpHeaders < Hash
     HTTP_PREFIX = /^HTTP_/
     RACK_HTTP_HEADER_ID = /#{HTTP_PREFIX.source}[A-Z_]+$/
